seperate teardown logic and fix pathing for autoload
